Adjust to new GitHub header
The box-shadow will get applied to the header element now. As a
fallback, if the old header style gets served, the repository-container-
header is used. Also, we now detect if a repository is public or not by
inspecting the octolytics- dimension-repository_public meta header. Both
changes should make the extension a bit more robust.